requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
discord.py==2.0.1
emoji-country-flag==1.3.1
Flask==2.2.2
frozenlist==1.3.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
